backend/pyproject.toml
[project]
name = "textpress-backend"
version = "0.1.0"
description = "Textpress hosted MVP backend"
requires-python = ">=3.11,<4.0"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.7.0",
  "SQLAlchemy>=2.0.30",
  "psycopg2-binary>=2.9.9",
  "textpress>=0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
package = false

#[tool.uv.sources]
# Use the published package by default; uncomment below to develop against local textpress
# textpress = { path = "..", editable = true }


